+++ python/lsst/ts/m2gui/__init__.py
@@ -37,7 +37,15 @@ from .enums import (
     TemperatureGroup,
     DisplacementSensorDirection,
 )
-from .utils import set_button, create_label, create_group_box, get_tol
+from .utils import (
+    NUM_ACTUATOR,
+    set_button,
+    create_label,
+    create_group_box,
+    create_table,
+    get_tol,
+    get_num_actuator_ring,
+)
 from .signals import (
     SignalControl,
     SignalMessage,
@@ -46,9 +54,11 @@ from .signals import (
     SignalLimitSwitch,
     SignalConfig,
     SignalUtility,
+    SignalDetailedForce,
 )
 from .log_window_handler import LogWindowHandler
 from .fault_manager import FaultManager
+from .actuator_force import ActuatorForce
 from .utility_monitor import UtilityMonitor
 from .config import Config
 from .model import Model